{Update} Grid Generator Script
Replace the camera at the center of the generated custom grid.
Change the orthographic size of the camera to half of the height of the texture2D used to generate the custom grid.
The width of the texture2D musn't be more than half of it's height to display everything at once.