{Update} Dialogue & DialogueManager Scripts
Dialogue: MyText class now has a string used to represent it's id and a second to choose where to jump after this text.

DialogueManager: minor changes to use the new system of string id/jumpTo instead of the old advanceDialogueBy int. Added a function to get the index of a text with it's id.