{Add} Level Prefab
Prefab composed of it's own Player, TileContainer, Inventory, GridGenerator and TouchInput.
This way to generate a new level, we just need to instantiate this prefab( and update the texture2D used by the GridGenerator) and deactivate ( SetActive(false) ) the previous level.

Some of these elements could be pulled out of this prefab (ex: GridGenerator). Don't know if I'll do it later.