{Update} GameManager, GridGenerator & Tile Script
GameManager: Added a Decor in the InfoLevel, pass the Decor to the GridGenerator

GridGenerator: Spawn the Decorations after the Tiles. The list of decoration needs to be at least the same length as the number of tile. Deleted the old tile offset (dk why I didn't deleted it earlier)

Tile: Commented out the line assigning the sprite order since now they are classed by their y value. Animation no longer playing when walking on a tile.