{Add} Buildings Script & Buildings Folders for Data & Prefabs
Buildings Script: Custom class to define the buildings of a level. While the Decor class takes a Vector2 correction for it's position (since instantiated at tile coordinates), the buildings directly take the coordinates they to be at (since they aren't linked to any tile).
Each Building needs to have it's own prefab (for future detection with the player)