{Update} GameManager & TouchInput Scripts
GameManager: Added SliderInfo into the InfoLevel class. Added a fonction to check if the correct year range is selected with the slider. Display the years from SliderInfo above the slider.
Also addeda bool and function to indicate if the player is interacting with the slider.

TouchInput: Checks the bool added in GameManager to avoid moving if the player is interacting with the slider instead of doing a sliding move.