{Update} Dialogue, DialogueCharacter & DialogueManager Scripts
Dialogue: Added 3 bools to MyText class. these bools "animate" the dialogue by stating if the professor is first alone and if other characters are entering leaving the stage.
Dialogue class now needs the first character other than the prof that appears in the dialogue (to avoid showing the wrong character (last from previous dialogue) at the start of a new dialogue)

DialogueCharacter: Added two Coroutines for entering and leaving the stage.

DialogueManager: Hide other character if Prof starts alone. Transformed NextText function to a Coroutine to handle entering/leaving the stage animations. Added a wrapper function (NextTextButton) to interact with the DialogueUI's buttons. Added a function to assign it's identity to the other character.