{Add} Characters Folder, Voxel Models, Animation & Animation Controller
Added the 3D model of the player and the professor along with their animations. Created an Animation controller to switch between the walking, waving1 and waving2 animations. Ignore "Nouveau Walking" animation, it's just to get a working model for the prof. Idk why but (maybe because they were in obj/mtl format) the model for the student and prof wouldn't animate but another model for the student that was inside waving2 is working. So I asked for another prof model inside an animation to get a working one.