{Update}/{Fix} Building, Player & GameManager Scripts
During the new level animation, if the player was on a building, the building wouldn't appear at the right place when the animation finishes (bad z axis). To fix that, buildings now ignore trigger enter if the animation is playing. after the animation, we deactivate and reactivate the player's hitbox to trigger the building for real this time.